pyproject.toml
[project]
name = "octo-travel-synthetic-api-simulator"
version = "0.1.0"
description = "Synthetic API simulator for OCTO travel products"
authors = [{ name = "Stefano Emilio Campanini", email = "stefano.campanini@gmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.28.0",
    "opentelemetry-api>=1.28.0",
]

[project.optional-dependencies]
test = []
dev = [
    "mypy>=1.13.0",
    "ruff>=0.8.0"
]

[project.scripts]
otas = "cli:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = ["src"]
package-dir = "src"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "hypothesis>=6.100.0",
    "httpx>=0.27.0",
]
